+++ tests/datamodules/test_datamodule_ddp.py
@@ -1,13 +1,14 @@
 import multiprocessing
-from argparse import Namespace
-import time
 import os
+import time
+from argparse import Namespace
 
 import pytest
 import pytorch_lightning as pl
 import torch
-from .test_utils import SimpleTransformerLikeModel, ExampleDataModule
 from transformers import BertTokenizer
+
+from .test_utils import ExampleDataModule, SimpleTransformerLikeModel
 
 n_cpus = multiprocessing.cpu_count()
 
@@ -15,27 +16,17 @@ n_cpus = multiprocessing.cpu_count()
 # Test iter dataset work correctly with dp
 @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
 @pytest.mark.parametrize(
-    ["ds_type", "num_workers", "distributed_backend", "gpus", "epochs", "dataset_idx"],
+    ["num_workers", "distributed_backend", "gpus", "epochs", "dataset_idx"],
     [
-
-    # ITER dataset
     # num_workers with ddp
-        ['iter', 0, 'ddp', 2, 1, 1],
-        ['iter', 1, 'ddp', 2, 2, 1],
-        ['iter', 2, 'ddp', 2, 2, 1],
-        ['iter', 0, 'ddp', 2, 1, 2],
-        ['iter', n_cpus, 'ddp', 2, 5, 2],
-
-    # MAP dataset
-    # num_workers with ddp
-        ['map', 0, 'ddp', 2, 2, 1],
-        ['map', 1, 'ddp', 2, 2, 1],
-        ['map', 2, 'ddp', 2, 2, 1],
-        ['map', 0, 'ddp', 2, 1, 2],
-        ['map', n_cpus, 'ddp', 2, 5, 3],
+        [0, 'ddp', 2, 2, 1],
+        [1, 'ddp', 2, 2, 1],
+        [2, 'ddp', 2, 2, 1],
+        [0, 'ddp', 2, 1, 2],
+        [n_cpus, 'ddp', 2, 5, 3],
     ]
 )
-def test_datamodule_gpu_ddp_only(ds_type, num_workers, distributed_backend, gpus, epochs, dataset_idx):
+def test_datamodule_gpu_ddp_only(num_workers, distributed_backend, gpus, epochs, dataset_idx):
 
     os.environ["CUDA_VISIBLE_DEVICES"] = "0,1"
     time.sleep(2)    # sleep for 5 second to be sure area is clean
@@ -51,7 +42,6 @@ def test_datamodule_gpu_ddp_only(ds_type, num_workers, distributed_backend, gpus
         max_steps=None,
         max_sequence_length=10,
         gpus=gpus,
-        iterable_datasets=ds_type == 'iter',
         accelerator=distributed_backend,
         skip_in_training=None
     )