+++ tests/datamodules/test_utils.py
@@ -1,11 +1,13 @@
 import torch
-from transformers.models.bert.modeling_bert import (BertConfig, BertForSequenceClassification)
 from transformers import AdamW
+from transformers.models.bert.modeling_bert import BertConfig, BertForSequenceClassification
+
 from transformers_lightning.adapters.csv_adapter import CSVAdapter
-from transformers_lightning import datamodules, models
+from transformers_lightning.datamodules import AdaptersDataModule
+from transformers_lightning.models import TransformersModel
 
 
-class SimpleTransformerLikeModel(models.TransformersModel):
+class SimpleTransformerLikeModel(TransformersModel):
 
     def __init__(self, hparams, do_ids_check=True):
         super().__init__(hparams)
@@ -101,7 +103,7 @@ class ExampleAdapter(CSVAdapter):
         return res
 
 
-class ExampleDataModule(datamodules.SuperDataModule):
+class ExampleDataModule(AdaptersDataModule):
 
     def __init__(self, *args, test_number=1, tokenizer=None, **kwargs):
         super().__init__(*args, **kwargs)