+++ transformers_lightning/adapters/csv_adapter.py
@@ -1,18 +1,20 @@
 import csv
+from argparse import Namespace
 from typing import Iterable
 
-from transformers_lightning.adapters import FileAdapter
+from transformers_lightning.adapters.file_adapter import FileAdapter
 from transformers_lightning.utils import strip_lines
 
 
 class CSVAdapter(FileAdapter):
     r"""
-    An Adapter to load data from a file. Only adds a `filepath` parameter.
-    It is still abstract and need to be subclassed.
+    An Adapter to load data from a CSV file. Only adds a `filepath` parameter.
     """
 
-    def __init__(self, *args, delimiter="\t", quoting=csv.QUOTE_MINIMAL, quotechar='"', **kwargs):
-        super().__init__(*args)
+    def __init__(
+        self, hparams: Namespace, filepath: str, delimiter="\t", quoting=csv.QUOTE_MINIMAL, quotechar='"', **kwargs
+    ):
+        super().__init__(hparams, filepath)
         self.csv_kwargs = {'delimiter': delimiter, 'quoting': quoting, 'quotechar': quotechar, **kwargs}
 
     def __iter__(self) -> Iterable: