+++ transformers_lightning/datamodules/__init__.py
@@ -1,1 +1,3 @@
-from transformers_lightning.datamodules.superdatamodule import SuperDataModule
+from transformers_lightning.datamodules.adapter_datamodule import AdaptersDataModule
+from transformers_lightning.datamodules.compressed_datamodule import CompressedDataModule
+from transformers_lightning.datamodules.super_datamodule import SuperDataModule