+++ transformers_lightning/datasets/__init__.py
@@ -1,5 +1,5 @@
 # triples of name as key, yaml config file and data file as values
-from transformers_lightning.datasets.superdataset import SuperTransformersDataset
-from transformers_lightning.datasets.map_dataset import TransformersMapDataset
-from transformers_lightning.datasets.iterable_dataset import TransformersIterableDataset
+from transformers_lightning.datasets.compressed_dataset import CompressedDataset
+from transformers_lightning.datasets.map_dataset import MapDataset
 from transformers_lightning.datasets.stack_dataset import StackDataset
+from transformers_lightning.datasets.super_dataset import SuperDataset